add concurrency asynchronous option

+++ async/src/main/java/fish/payara/codingdojo/async/resources/model/Data.java
@@ -6,5 +6,5 @@ import java.io.Serializable;
  *
  * @author Petr Aubrecht <aubrecht@asoftware.cz>
  */
-public record Data(long sourceId, long time, int position, int temperature, String label) implements Serializable {
+public record Data(long sourceId, double time, int position, int temperature, String label) implements Serializable {
 }
